feat(canvas): delete files that are not used and organize files
