--- src/Home.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Home = () => {
-  return (
-    <div className="Home">
-      <h1>Home Page</h1>
-    </div>
-  );
-};
-
-export default Home;
