--- src/pages/ImageGallery.js
@@ -1,18 +1,0 @@
-import React from "react";
-import useFirestore from "../hooks/useFirestore";
-import ImageGrid from "../components/ImageGrid";
-
-function ImagePage() {
-  const { docs } = useFirestore("images");
-  const data = useFirestore("images");
-  console.log(data);
-  return (
-    <div>
-      <h2>Images</h2>
-
-      <ImageGrid data={docs} />
-    </div>
-  );
-}
-
-export default ImagePage;
